+++ packages/cspell-lib/src/textValidation/textValidator.test.ts
@@ -9,7 +9,7 @@ import { FreqCounter } from '../util/FreqCounter';
 import * as Text from '../util/text';
 import { _testMethods, calcTextInclusionRanges, validateText } from './textValidator';
 import type { ValidationOptions } from './ValidationTypes';
-import { settingsToValidateOptions } from './validator';
+import { settingsToValidateOptions } from './settingsToValidateOptions';
 
 function sToV(settings: CSpellUserSettings) {
     return settingsToValidateOptions(finalizeSettings(settings));
